apply changes

+++ app/Http/Controllers/TagsController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\TagRequest;
 use App\Models\Tag;
+use Faker\Extension\GeneratorAwareExtensionTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 
 class TagsController extends Controller
@@ -18,18 +20,26 @@ class TagsController extends Controller
 
     public function index()
     {
-
-        $tags = Tag::all();//return all tags are store in the table as collection object can use as array contain object of tag
+        if (!Gate::allows('tags.view')) {
+            abort(403);
+        }
+        $tags = Tag::paginate();//return all tags are store in the table as collection object can use as array contain object of tag
         return view('tags.index', ['title' => 'Tag List', 'tags' => $tags]);
     }
 
     public function create()
     {
+        if (Gate::denies('tags.create')) {
+            abort(403);
+        }
         return view('tags.create', ['title' => 'Create New Tag']);
     }
 
     public function store(TagRequest $request)
     {
+        if (Gate::denies('tags.create')) {
+            abort(403);
+        }
 //        $this->validateRequest($request);
 //        $request->all(); //return array contain token & data was sent from form
         $request->merge(['slug' => Str::slug($request->name)]);
@@ -39,12 +49,14 @@ class TagsController extends Controller
 
     public function edit($id)
     {
+        Gate::authorize('tags.edit');
         $tag = Tag::findOrfail($id);
         return view('tags.edit', ['tag' => $tag, 'title' => 'Edit Tag']);
     }
 
     public function update(TagRequest $request, $id)
     {
+        Gate::authorize('tags.edit');
 //        $this->validateRequest($request, $id);
         $tag = Tag::findOrfail($id);
         $tag->update(['name' => $request->name, 'slug' => Str::slug($request->name)]);
@@ -53,6 +65,7 @@ class TagsController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('tags.delete');
         Tag::destroy($id);
         return redirect('/tags')->with('success', 'Tag deleted!');
